% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bff.R
\name{bff}
\alias{bff}
\title{Find features most associated with cluster membership}
\usage{
bff(loadings, features, num_best, ...)
}
\arguments{
\item{loadings}{An \code{n} by \code{k} matrix of weights that indicates how
important that ith user is to the jth cluster, i.e., the \code{Z} or \code{Y}
matrix calculated by \code{\link[=vsp]{vsp()}}.}

\item{features}{An \code{n} by \code{d} matrix of features measured for each
node in the network.}

\item{num_best}{An integer indicating how many of the top features
for differentiating between loadings you want.}
}
\value{
An \code{n} by \code{k} matrix whose \verb{[i, j]} entry is the
ith "most important" feature for cluster j.
}
\description{
Find features most associated with cluster membership
}
