% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidygraph.R
\name{bind_pca_factors}
\alias{bind_pca_factors}
\title{Add PCA factor loadings to node table of tidygraph}
\usage{
bind_pca_factors(graph, fa = NULL, ...)
}
\arguments{
\item{graph}{A [tidygraph::tbl_graph] object.}

\item{fa}{Optionally, a [vsp] object to extract varimax loadings from. If
you do not passed a [vsp] object, one will be created.}

\item{...}{Arguments passed on to \code{vsp}
\describe{
  \item{x}{Either a graph adjacency matrix, [igraph::igraph] or
[tidygraph::tbl_graph]. If `x` is a [matrix] or [Matrix::Matrix]
then `x[i, j]` should correspond to the edge going from node `i`
to node `j`.}
  \item{k}{The number of factors to calculate.}
  \item{center}{Should the adjacency matrix be row *and* column centered?
Defaults to `TRUE`.}
  \item{normalize}{Should the graph laplacian be used instead of the
raw adjacency matrix? Defaults to `TRUE`. If `center = TRUE`, `A` will
first be centered and then normalized.}
  \item{tau_row}{Row regularization term. Default is `NULL`, in which case
we use the row degree. Ignored when `normalize = FALSE`.}
  \item{tau_col}{Column regularization term. Default is `NULL`, in which case
we use the column degree. Ignored when `normalize = FALSE`.}
}}
}
\value{
The same `graph` object with columns `factor1`, ..., `factor{k}`
 in the table of node information.
}
\description{
Add PCA factor loadings to node table of tidygraph
}
