% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localization.R
\name{cumulative_participation}
\alias{cumulative_participation}
\title{Calculate cumulative participation of a set of singular vectors.}
\usage{
cumulative_participation(U)
}
\arguments{
\item{U}{A matrix of singular vectors.}
}
\value{
A scalar numeric value representing the cumulative participation.
}
\description{
Calculate cumulative participation of a set of singular vectors.
}
