% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{get_localization}
\alias{get_localization}
\title{Get the localization measure for each factor}
\usage{
get_localization(fa)
}
\arguments{
\item{fa}{A [vsp][vsp-object] object created by a call to [vsp()].}
}
\value{
TODO
}
\description{
Get the localization measure for each factor
}
\examples{

set.seed(27)
M <- sign(rsparsematrix(12, 12, nnz = 40))^2

fa <- vsp(M, k = 4)

get_localization(fa)

}
\seealso{
Other diagnostics: \code{\link{plot_localization}}
}
\concept{diagnostics}
