% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{get_varimax_y}
\alias{get_varimax_y}
\title{Title}
\usage{
get_varimax_y(x, factors = 1:x$rank, ...)
}
\arguments{
\item{x}{TODO}

\item{factors}{TODO}

\item{...}{TODO}
}
\value{
TODO
}
\description{
Title
}
