% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{get_z_clusters}
\alias{get_z_clusters}
\title{Assign nodes to clusters via a rough heuristic}
\usage{
get_z_clusters(x, ...)
}
\arguments{
\item{x}{A factor analysis object.}

\item{...}{Ignored.}
}
\description{
This is a heuristic that assigns nodes clusters based on highest
absolute value of loadings. Each node is assigned to a single cluster.#'
}
