% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{get_z_hubs}
\alias{get_z_hubs}
\title{Title}
\usage{
get_z_hubs(fa, hubs_per_factor = 10, factors = 1:fa$rank)
}
\arguments{
\item{fa}{TODO}

\item{hubs_per_factor}{TODO}

\item{factors}{TODO}
}
\value{
TODO
}
\description{
Title
}
