% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localization.R
\name{ipr}
\alias{ipr}
\title{Calculate the inverse participation ratio (IPR) for a vector.}
\usage{
ipr(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A scalar numeric value representing the IPR.
}
\description{
Calculate the inverse participation ratio (IPR) for a vector.
}
