% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{left_padded_sequence}
\alias{left_padded_sequence}
\title{Title}
\usage{
left_padded_sequence(x)
}
\arguments{
\item{x}{A vector of integers, small enough the scientific}
}
\value{
TODO
}
\description{
Title
}
\keyword{internal}
