% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_skew_positive}
\alias{make_skew_positive}
\title{Make factors columnwise skew positive}
\usage{
make_skew_positive(fa)
}
\arguments{
\item{fa}{A \code{\link[=fa_like]{fa_like()}} object.}
}
\value{
A new \code{\link[=fa_like]{fa_like()}} object where the columns
of \code{Z} and \code{Y} has positive skew, that is otherwise
equivalent to the original object.
}
\description{
Given a factor analysis like object, flip
signs so that columns of \code{Z} and \code{Y} are
skew positive. Note that this also causes
corresponding sign flips in \code{B}. This
helps with interpretability of factors.
}
\keyword{internal}
