% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{match_column_order}
\alias{match_column_order}
\title{Match co-factors across Z and Y to the extent possible}
\usage{
match_column_order(fa)
}
\arguments{
\item{fa}{A \code{\link[=fa_like]{fa_like()}} object.}
}
\value{
A new \code{\link[=fa_like]{fa_like()}} object where the columns
of \code{Z} and \code{Y} have been re-ordered to correspond with each other,
by making \code{B} as diagonally dominant as possible.
}
\description{
Match co-factors across Z and Y to the extent possible
}
\keyword{internal}
