% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localization.R
\name{plot_cumulative_curves}
\alias{plot_cumulative_curves}
\title{Plot cumulative participation curves.}
\usage{
plot_cumulative_curves(localization)
}
\arguments{
\item{localization}{A \code{localization_stats} object.}
}
\value{
A \code{ggplot2} object.
}
\description{
Plot cumulative participation curves.
}
\examples{
# See localization_statistics for examples
}
