% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_mixing_matrix}
\alias{plot_mixing_matrix}
\title{Plot the mixing matrix B}
\usage{
plot_mixing_matrix(fa)
}
\arguments{
\item{fa}{A \code{\link[=vsp_fa]{vsp_fa()}} object.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with one row for each node, and one column
containing each of the requested factor or singular vector, plus
an additional \code{id} column.
}
\description{
Plot the mixing matrix B
}
