% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_varimax_z_pairs}
\alias{plot_varimax_z_pairs}
\title{Title}
\usage{
plot_varimax_z_pairs(fa, factors = 1:min(5, fa$rank), ...)
}
\arguments{
\item{fa}{TODO}

\item{factors}{TODO}

\item{...}{TODO}
}
\value{
TOOD
}
\description{
Title
}
