% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regularize.R
\name{regularized_laplacian}
\alias{regularized_laplacian}
\alias{regularized_laplacian.igraph}
\title{Get the regularized graph laplacian of a graph}
\usage{
regularized_laplacian(graph, ...)

\method{regularized_laplacian}{igraph}(graph, ..., tau_row = NULL,
  tau_col = NULL, sparse = TRUE)
}
\arguments{
\item{graph}{Either a graph adjacency matrix, [igraph::igraph] or
[tidygraph::tbl_graph]. If `x` is a [matrix] or [Matrix::Matrix]
then `x[i, j]` should correspond to the edge going from node `i`
to node `j`.}

\item{...}{Ignored.}

\item{tau_row}{Row regularization term. Default is `NULL`, in which case
we use the mean row degree. Ignored when `normalize = FALSE`.}

\item{tau_col}{Column regularization term. Default is `NULL`, in which case
we use the mean column degree. Ignored when `normalize = FALSE`.}
}
\value{
The regularized graph Laplacian as a [Matrix::Matrix()].
}
\description{
Get the regularized graph laplacian of a graph
}
