% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{new_vsp}
\alias{new_vsp}
\title{Create a new `vsp` object}
\usage{
new_vsp(U, d, V, Z, B, Y, center, normalize, k, tau_list)
}
\description{
Users should use `vsp`, this is a low-level internal constructor.
}
