% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsp-package.R
\docType{package}
\name{vsp-package}
\alias{vsp}
\alias{vsp-package}
\title{vsp: Vintage Sparse PCA for Non-Parametric Factor Analysis}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\author{
\strong{Maintainer}: Alex Hayes \email{alexpghayes@gmail.com} (0000-0002-4985-5160)

Authors:
\itemize{
  \item Karl Rohe \email{karlrohe@stat.wisc.edu}
}

}
\keyword{internal}
