% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsp.R
\name{vsp}
\alias{vsp}
\title{Non-Parametric Factor Analysis via Vintage Sparse PCA}
\usage{
vsp(A, k = 5, tau_row = NULL, tau_col = NULL, normalize = TRUE,
  center = FALSE)
}
\arguments{
\item{A}{A [matrix] or [Matrix::Matrix] object.}

\item{k}{The number of factors to calculate.}

\item{tau_row}{Row regularization term. Defaults is `NULL`, in which case
we use the row degree.}

\item{tau_col}{Column regularization term. Defaults is `NULL`, in which case
we use the column degree.}

\item{normalize}{Should the graph laplacian be used instead of the
raw adjacency matrix? Defaults to `TRUE`.}
}
\value{
An object of class `vsp`
}
\description{
This code implements regularized spectral clustering with varimax.
allows for symmetric, directed, and bipartite (i.e. rectangular A).
}
\details{
Uses `RSpectra` for matrix computations. These are best-in-class
 implementations of sparse matrix eigendecompositions. The implementations
 are not parallel, however.

 TODO: how the normalization is done, how the centering is done
}
