% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidygraph.R
\name{bind_varimax_z}
\alias{bind_varimax_z}
\title{Add varimax factor loadings to node table of tidygraph}
\usage{
bind_varimax_z(graph, fa = NULL, ...)
}
\arguments{
\item{graph}{A \link[tidygraph:tbl_graph]{tidygraph::tbl_graph} object.}

\item{fa}{Optionally, a \link{vsp} object to extract varimax loadings from. If
you do not passed a \link{vsp} object, one will be created.}

\item{...}{
  Arguments passed on to \code{\link[=vsp]{vsp}}
  \describe{
    \item{\code{x}}{Either a graph adjacency matrix, \link[igraph:aaa-igraph-package]{igraph::igraph} or
\link[tidygraph:tbl_graph]{tidygraph::tbl_graph}. If \code{x} is a \link{matrix} or \link[Matrix:Matrix]{Matrix::Matrix}
then \code{x[i, j]} should correspond to the edge going from node \code{i}
to node \code{j}.}
    \item{\code{k}}{The number of factors to calculate.}
    \item{\code{center}}{Should the adjacency matrix be row \emph{and} column centered?
Defaults to \code{TRUE}.}
    \item{\code{normalize}}{Should the graph laplacian be used instead of the
raw adjacency matrix? Defaults to \code{TRUE}. If \code{center = TRUE}, \code{A} will
first be centered and then normalized.}
    \item{\code{tau_row}}{Row regularization term. Default is \code{NULL}, in which case
we use the row degree. Ignored when \code{normalize = FALSE}.}
    \item{\code{tau_col}}{Column regularization term. Default is \code{NULL}, in which case
we use the column degree. Ignored when \code{normalize = FALSE}.}
  }}
}
\value{
The same \code{graph} object with columns \code{factor1}, ..., \verb{factor\{k\}}
in the table of node information.
}
\description{
Add varimax factor loadings to node table of tidygraph
}
