% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localization.R
\name{localization_statistics}
\alias{localization_statistics}
\title{Title}
\usage{
localization_statistics(
  graph,
  max_rank,
  ...,
  tau_min = 10^-2,
  tau_max = 10^4,
  num_tau = 50
)
}
\arguments{
\item{num_tau}{}
}
\description{
Title
}
\examples{

library(igraphdata)
library(furrr)

data(karate, package = "igraphdata")

plan(multicore, workers = 10)

# karate is undirected, enron is directed

stats <- localization_statistics(karate, max_rank = 15, num_tau = 200)

plot_cumulative_curves(stats)
plot_ipr_curves(stats)

}
