% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{screeplot.vsp_fa}
\alias{screeplot.vsp_fa}
\title{Create a screeplot from a factor analysis object}
\usage{
\method{screeplot}{vsp_fa}(x, ...)
}
\arguments{
\item{x}{A \code{\link[=vsp_fa]{vsp_fa()}} object.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with one row for each node, and one column
containing each of the requested factor or singular vector, plus
an additional \code{id} column.
}
\description{
Create a screeplot from a factor analysis object
}
