% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsp-package.R
\docType{package}
\name{vsp-package}
\alias{vsp-package}
\title{vsp: Vintage Sparse PCA for Semi-Parametric Factor Analysis}
\description{
Provides fast spectral estimation of latent factors in random dot product graphs using the vsp estimator. Under mild assumptions, the vsp estimator is consistent for (degree-corrected) stochastic blockmodels, (degree-corrected) mixed-membership stochastic blockmodels, and degree-corrected overlapping stochastic blockmodels.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rohelab.github.io/vsp/}
  \item \url{https://github.com/RoheLab/vsp}
  \item Report bugs at \url{https://github.com/RoheLab/vsp/issues}
}

}
\author{
\strong{Maintainer}: Alex Hayes \email{alexpghayes@gmail.com} (\href{https://orcid.org/0000-0002-4985-5160}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Karl Rohe \email{karlrohe@stat.wisc.edu}
  \item Muzhe Zeng \email{mzeng6@wisc.edu}
  \item Fan Chen
}

}
\keyword{internal}
