% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsp.R
\name{vsp}
\alias{vsp}
\title{Non-Parametric Factor Analysis via Vintage Sparse PCA}
\usage{
vsp(A, k = 5, tau = NULL, normalize = TRUE, center = TRUE,
  sym = NULL)
}
\arguments{
\item{A}{A [matrix] or [Matrix::Matrix] object.}

\item{k}{is the number of clusters.  must be same for rows and columns.}

\item{tau}{is regularization parameter.  it can be a 2 vectors.
Default is pretty good.}

\item{normalize}{Should the adjacency matrix be normalized?}

\item{center}{Should the adjacency matrix `A` be row and column centered?}

\item{sym}{Set to `FALSE` for a two-way analysis.}
}
\value{
An object of class `vsp`, which is a list with elements:

  - `V`: only returned in the rectangular case
  - `U`: TODO
  - `Z`: TODO
  - `B`: TODO
  - `scree`: TODO
  - `Y`: TODO
}
\description{
This code implements regularized spectral clustering with varimax.
allows for symmetric, directed, and bipartite (i.e. rectangular A).
}
\details{
The `irlba` package is used for non-centered calculations, which
 requires calculating a partial SVD. Calculations for centered problems
 use `rARPACK`.

 TODO: how the normalization is done, how the centering is done
}
