% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsp.R
\name{vsp}
\alias{vsp}
\alias{vsp.default}
\alias{vsp.igraph}
\title{Non-Parametric Factor Analysis via Vintage Sparse PCA}
\usage{
vsp(x, k = 5, center = TRUE, normalize = TRUE, tau_row = NULL,
  tau_col = NULL)

\method{vsp}{default}(x, k = 5, center = TRUE, normalize = TRUE,
  tau_row = NULL, tau_col = NULL)

\method{vsp}{igraph}(x, k = 5, center = TRUE, normalize = TRUE,
  tau_row = NULL, tau_col = NULL)
}
\arguments{
\item{x}{Either a graph adjacency matrix, [igraph::igraph] or
[tidygraph::tbl_graph]. If `x` is a [matrix] or [Matrix::Matrix]
then `x[i, j]` should correspond to the edge going from node `i`
to node `j`.}

\item{k}{The number of factors to calculate.}

\item{center}{Should the adjacency matrix be row *and* column centered?
Defaults to `TRUE`.}

\item{normalize}{Should the graph laplacian be used instead of the
raw adjacency matrix? Defaults to `TRUE`. If `center = TRUE`, `A` will
first be centered and then normalized.}

\item{tau_row}{Row regularization term. Default is `NULL`, in which case
we use the row degree. Ignored when `normalize = FALSE`.}

\item{tau_col}{Column regularization term. Default is `NULL`, in which case
we use the column degree. Ignored when `normalize = FALSE`.}
}
\value{
An object of class `vsp`. TODO: Details
}
\description{
This code implements TODO.
}
\details{
Sparse SVDs use `RSpectra` for performance.
}
