% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsp.R
\name{vsp}
\alias{vsp}
\alias{vsp.default}
\alias{vsp.matrix}
\alias{vsp.Matrix}
\alias{vsp.dgCMatrix}
\alias{vsp.igraph}
\title{Non-Parametric Factor Analysis via Vintage Sparse PCA}
\usage{
vsp(x, rank, ...)

\method{vsp}{default}(x, rank, ...)

\method{vsp}{matrix}(
  x,
  rank,
  ...,
  center = FALSE,
  recenter = FALSE,
  scale = TRUE,
  rescale = scale,
  tau_row = NULL,
  tau_col = NULL
)

\method{vsp}{Matrix}(
  x,
  rank,
  ...,
  center = FALSE,
  recenter = FALSE,
  scale = TRUE,
  rescale = scale,
  tau_row = NULL,
  tau_col = NULL
)

\method{vsp}{dgCMatrix}(
  x,
  rank,
  ...,
  center = FALSE,
  recenter = FALSE,
  scale = TRUE,
  rescale = scale,
  tau_row = NULL,
  tau_col = NULL
)

\method{vsp}{igraph}(x, rank, ..., attr = NULL)
}
\arguments{
\item{x}{Either a graph adjacency matrix, \link[igraph:aaa-igraph-package]{igraph::igraph} or
\link[tidygraph:tbl_graph]{tidygraph::tbl_graph}. If \code{x} is a \link{matrix} or \link[Matrix:Matrix]{Matrix::Matrix}
then \code{x[i, j]} should correspond to the edge going from node \code{i}
to node \code{j}.}

\item{...}{Ignored.}

\item{center}{Should the adjacency matrix be row \emph{and} column centered?
Defaults to \code{TRUE}.}

\item{tau_row}{Row regularization term. Default is \code{NULL}, in which case
we use the row degree. Ignored when \code{normalize = FALSE}.}

\item{tau_col}{Column regularization term. Default is \code{NULL}, in which case
we use the column degree. Ignored when \code{normalize = FALSE}.}

\item{k}{The number of factors to calculate.}

\item{normalize}{Should the graph laplacian be used instead of the
raw adjacency matrix? Defaults to \code{TRUE}. If \code{center = TRUE}, \code{A} will
first be centered and then normalized.}
}
\value{
An object of class \code{vsp}. TODO: Details
}
\description{
This code implements TODO.
}
\details{
Sparse SVDs use \code{RSpectra} for performance.
}
\examples{

library(LRMF3)

vsp(ml100k, rank = 5, scale = TRUE)
vsp(ml100k, rank = 5, rescale = FALSE)
vsp(ml100k, rank = 5)


}
