% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsp.R
\name{vsp.svd_like}
\alias{vsp.svd_like}
\title{Perform varimax rotation on a low rank matrix factorization}
\usage{
\method{vsp}{svd_like}(
  x,
  rank,
  ...,
  centerer = NULL,
  scaler = NULL,
  recenter = FALSE,
  rescale = FALSE,
  kaiser_normalize_u = FALSE,
  kaiser_normalize_v = FALSE,
  rownames = NULL,
  colnames = NULL
)
}
\arguments{
\item{x}{Either a graph adjacency matrix, \link[igraph:aaa-igraph-package]{igraph::igraph} or
\link[tidygraph:tbl_graph]{tidygraph::tbl_graph}. If \code{x} is a \link{matrix} or \link[Matrix:Matrix]{Matrix::Matrix}
then \code{x[i, j]} should correspond to the edge going from node \code{i}
to node \code{j}.}

\item{rank}{The number of factors to calculate.}

\item{...}{Ignored.}

\item{centerer}{TODO}

\item{scaler}{TODO}

\item{recenter}{Should the varimax factors be re-centered around the
original factor means? Only used when \code{center = TRUE}, defaults to \code{FALSE}.}

\item{rescale}{Should the regularized graph laplacian be used instead of the
raw adjacency matrix? Defaults to \code{TRUE}. If \code{center = TRUE}, \code{A} will
first be centered and then normalized.}

\item{kaiser_normalize_u}{Whether or not to use Kaiser normalization
when rotating the left singular vectors \code{U}. Defaults to \code{FALSE}.}

\item{kaiser_normalize_v}{Whether or not to use Kaiser normalization
when rotating the right singular vectors \code{V}. Defaults to \code{FALSE}.}

\item{rownames}{Character vector of row names of \code{x}. These row names
are propagated into the row names of the \code{U} and \code{Z}. Defaults
to \code{NULL}.}

\item{colnames}{Character vector of column names of \code{x}. These column names
are propagated into the row names of the \code{V} and \code{Y}. Defaults
to \code{NULL}.}
}
\description{
Perform varimax rotation on a low rank matrix factorization
}
\examples{

library(fastadi)

mf <- adaptive_impute(ml100k, rank = 20, max_iter = 5)
fa <- vsp(mf)

}
