% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsp.R
\name{vsp.svd_like}
\alias{vsp.svd_like}
\title{Perform varimax rotation on a low rank matrix factorization}
\usage{
\method{vsp}{svd_like}(
  x,
  rank,
  ...,
  centerer = NULL,
  scaler = NULL,
  recenter = FALSE,
  rescale = TRUE
)
}
\arguments{
\item{x}{}

\item{scaler}{}
}
\description{
Perform varimax rotation on a low rank matrix factorization
}
\examples{

library(fastadi)

mf <- adaptive_impute(ml100k, rank = 20, max_iter = 5)
fa <- vsp(mf)

}
