% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{vsp_fa}
\alias{vsp_fa}
\title{Create a vintage sparse factor analysis object}
\usage{
vsp_fa(u, d, v, Z, B, Y, transformers, R_U, R_V)
}
\arguments{
\item{u}{A \emph{matrix} "left singular-ish" vectors.}

\item{d}{A \emph{vector} of "singular-ish" values.}

\item{v}{A \emph{matrix} of "right singular-ish" vectors.}

\item{Z}{TODO}

\item{B}{TODO}

\item{Y}{TODO}

\item{transformers}{TODO}
}
\value{
An \code{adaptive_imputation} object.
}
\description{
\code{adaptive_imputation} objects are a subclass of
\code{\link[LRMF3:svd_like]{LRMF3::svd_like()}}, with an additional field \code{alpha}.
}
